requirements.txt
langchain
python-dotenv
langchain_community
pypdf 
arxiv
pymupdf
langchain-text-splitters
langchain_huggingface
sentence_transformers
streamlit
langchain_core
chroma
langchain_chroma